Ubuntu: setup networking in guest
We create a simple netplan setup in charge of getting an IP from the
qemu host through DHCP.

With the current config, the host qemu acts as a gateway to the host
default route.

With that, we can reach internet from the VM when running outside of
the Nix sandbox (through the .driver and .driverInteractive passthru).