Introduce Fedora support
We'll support Fedora 38 and 39 for now. The two currently maintained
versions.

For some reason, the 9P Nix store mount works fine on the fedora
hosts, but aren't accessible from the systemd services. I investigated
this for a while and ended giving up on using the store from the
system services.

I know for sure this ain't related to:

- SELinux
- Systemd hardening

My gut instinct tells me this is a weird interaction beween 9P and
recent systemd. I checked the systemd bug tracker, no luck.

I removed all the store references from the backdoor and resizedisk
systemd services to work around that bug.