Remove VM name option
We currently do not support multi-machines setups. The name is just a
footgun: it ends up in a Python variable. Set it up to any value that
is not compatible with a Python variable and you'll end up with a
puzzling error message.

Let's remove it and replace it with a simple "vm".